(maint) Remove pdk/analytics require from spec_helper
