+++ lib/pdk/analytics/client/google_analytics.rb
@@ -1,3 +1,5 @@
+require 'pdk'
+
 module PDK
   module Analytics
     module Client
