+++ lib/pdk/analytics/util.rb
@@ -1,3 +1,5 @@
+require 'pdk'
+
 module PDK
   module Analytics
     module Util
