+++ lib/pdk/cli/util/command_redirector.rb
@@ -1,3 +1,4 @@
+require 'pdk'
 require 'tty-prompt'
 
 module PDK
