+++ lib/pdk/cli/util/interview.rb
@@ -1,4 +1,5 @@
 require 'tty-prompt'
+require 'pdk'
 
 module PDK
   module CLI
