+++ lib/pdk/cli/util/option_validator.rb
@@ -1,3 +1,5 @@
+require 'pdk'
+
 module PDK
   module CLI
     module Util
