+++ lib/pdk/util/filesystem.rb
@@ -1,3 +1,5 @@
+require 'pdk'
+
 module PDK
   module Util
     module Filesystem
